finish basic middleware structure
