+++ app/src/js/common/sync_middleware.ts
@@ -1,8 +1,54 @@
 import { Dispatch, Middleware, MiddlewareAPI } from 'redux'
-import { setStatusAfterConnect } from '../action/common'
 import * as types from '../action/types'
+import { State } from '../functional/types'
 import { ReduxState } from './configure_store'
+import { setupSession } from './session_setup'
 import { Synchronizer } from './synchronizer'
+
+/**
+ * Handles actions related to socket.io and synchronization
+ */
+function handleSyncAction  (
+  action: types.BaseAction, synchronizer: Synchronizer, state: State) {
+  switch (action.type) {
+    case types.REGISTER_SESSION:
+      const initialState = (action as types.RegisterSessionAction).initialState
+      synchronizer.finishRegistration(initialState,
+        initialState.task.config.autosave,
+        initialState.session.id,
+        initialState.task.config.bots)
+      break
+    case types.CONNECT:
+      synchronizer.sendConnectionMessage(state.session.id)
+      break
+    case types.DISCONNECT:
+      synchronizer.handleDisconnect()
+      break
+    case types.RECEIVE_BROADCAST:
+      const message = (action as types.ReceiveBroadcastAction).message
+      synchronizer.handleBroadcast(message)
+      break
+    case types.SAVE:
+      synchronizer.sendQueuedActions(
+        state.session.id, state.task.config.bots)
+      break
+  }
+}
+
+/**
+ * Handles autosaving of a normal action
+ */
+function handleNormalAction (
+  action: types.BaseAction, synchronizer: Synchronizer, state: State) {
+  const sessionId = state.session.id
+  const autosave = state.task.config.autosave
+  const bots = state.task.config.bots
+
+  if (sessionId === action.sessionId && !action.frontendOnly &&
+      !types.isSessionAction(action)) {
+    synchronizer.logAction(action, autosave, sessionId, bots)
+  }
+}
 
 export const makeSyncMiddleware = (synchronizer: Synchronizer) => {
   const syncMiddleware: Middleware<ReduxState> = (
@@ -10,29 +56,17 @@ export const makeSyncMiddleware = (synchronizer: Synchronizer) => {
 
     return (next: Dispatch) => (action: types.BaseAction) => {
       const state = getState().present
-      const sessionId = state.session.id
-      const autosave = state.task.config.autosave
 
-      // Handle socket events
       if (types.isSyncAction(action)) {
-        switch (action.type) {
-          case types.REGISTER_SESSION:
-          case types.CONNECT:
-            synchronizer.sendConnectionMessage(sessionId)
-            dispatch(setStatusAfterConnect())
-            break
-          case types.DISCONNECT:
-          case types.RECEIVE_BROADCAST:
-          default:
-            // throw error- should all be covered
-        }
+        // Handle socket events
+        handleSyncAction(action, synchronizer, state)
         return action
+      } else if (action.type === types.UPDATE_STATE) {
+        const returnValue = next(action)
+        setupSession()
+        return returnValue
       } else {
-        if (sessionId === action.sessionId && !action.frontendOnly &&
-          !types.isSessionAction(action)) {
-            // sync.queueAction
-        }
-
+        handleNormalAction(action, synchronizer, state)
         return next(action)
       }
     }
